+++ packages/ui/src/calendar/test/calendar.test.tsx
@@ -136,11 +136,6 @@ describe( 'Calendar', () => {
 
 		it( 'should show multiple months at once via the `numberOfMonths` prop', () => {
 			render( <Calendar numberOfMonths={ 2 } /> );
-			expect(
-				screen.getByRole( 'application', {
-					name: 'Date calendar, May 2025 and June 2025',
-				} )
-			).toBeVisible();
 
 			const grids = screen.getAllByRole( 'grid' );
 			expect( grids ).toHaveLength( 2 );