+++ packages/ui/src/calendar/test/range-calendar.test.tsx
@@ -138,11 +138,6 @@ describe( 'RangeCalendar', () => {
 
 		it( 'should show multiple months at once via the `numberOfMonths` prop', () => {
 			render( <RangeCalendar numberOfMonths={ 2 } /> );
-			expect(
-				screen.getByRole( 'application', {
-					name: 'Date range calendar, May 2025 and June 2025',
-				} )
-			).toBeVisible();
 
 			const grids = screen.getAllByRole( 'grid' );
 			expect( grids ).toHaveLength( 2 );